chore(scap): route getpid_global through scap_platform
