+++ userspace/libscap/engine/nodriver/nodriver.c
@@ -76,7 +76,6 @@ const struct scap_vtable scap_nodriver_engine = {
 	.get_n_devs = noop_get_n_devs,
 	.get_max_buf_used = noop_get_max_buf_used,
 	.get_threadlist = noop_get_threadlist,
-	.getpid_global = noop_getpid_global,
 	.get_api_version = NULL,
 	.get_schema_version = NULL,
 };
