+++ userspace/libscap/engine/noop/noop.c
@@ -106,11 +106,6 @@ int32_t noop_get_threadlist(struct scap_engine_handle engine, struct ppm_proclis
 	return SCAP_NOT_SUPPORTED;
 }
 
-int32_t noop_getpid_global(struct scap_engine_handle engine, int64_t* pid, char* error)
-{
-	return SCAP_NOT_SUPPORTED;
-}
-
 const struct scap_vtable scap_noop_engine = {
 	.name = "noop",
 	.mode = SCAP_MODE_NODRIVER,
@@ -129,7 +124,6 @@ const struct scap_vtable scap_noop_engine = {
 	.get_n_devs = noop_get_n_devs,
 	.get_max_buf_used = noop_get_max_buf_used,
 	.get_threadlist = noop_get_threadlist,
-	.getpid_global = noop_getpid_global,
 	.get_api_version = NULL,
 	.get_schema_version = NULL,
 };
