+++ userspace/libscap/engine/source_plugin/source_plugin.c
@@ -247,7 +247,6 @@ const struct scap_vtable scap_source_plugin_engine = {
 	.get_n_devs = noop_get_n_devs,
 	.get_max_buf_used = noop_get_max_buf_used,
 	.get_threadlist = noop_get_threadlist,
-	.getpid_global = noop_getpid_global,
 	.get_api_version = NULL,
 	.get_schema_version = NULL,
 };
