+++ userspace/libscap/engine/test_input/test_input.c
@@ -125,7 +125,6 @@ const struct scap_vtable scap_test_input_engine = {
 	.get_n_devs = noop_get_n_devs,
 	.get_max_buf_used = noop_get_max_buf_used,
 	.get_threadlist = NULL,
-	.getpid_global = noop_getpid_global,
 	.get_api_version = NULL,
 	.get_schema_version = NULL,
 };
