+++ userspace/libscap/linux/scap_linux_int.h
@@ -43,6 +43,7 @@ uint32_t scap_linux_get_device_by_mount_id(struct scap_platform* platform, const
 struct scap_threadinfo* scap_linux_proc_get(struct scap_platform* platform, struct scap_proclist* proclist, int64_t tid, bool scan_sockets);
 int32_t scap_linux_refresh_proc_table(struct scap_platform* platform, struct scap_proclist* proclist);
 bool scap_linux_is_thread_alive(struct scap_platform* platform, int64_t pid, int64_t tid, const char* comm);
+int32_t scap_linux_getpid_global(struct scap_platform* platform, int64_t *pid, char* error);
 
 // read all sockets and add them to the socket table hashed by their ino
 int32_t scap_fd_read_sockets(char* procdir, struct scap_ns_socket_list* sockets, char *error);
