+++ userspace/libscap/linux/scap_linux_platform.c
@@ -139,6 +139,7 @@ static const struct scap_platform_vtable scap_linux_platform = {
 	.get_proc = scap_linux_proc_get,
 	.refresh_proc_table = scap_linux_refresh_proc_table,
 	.is_thread_alive = scap_linux_is_thread_alive,
+	.get_global_pid = scap_linux_getpid_global,
 	.close_platform = scap_linux_close_platform,
 	.free_platform = scap_linux_free_platform,
 };
