+++ userspace/libscap/linux/scap_linux_platform.h
@@ -55,6 +55,15 @@ struct scap_linux_vtable {
 	 * PID namespace
 	 */
 	int32_t (*get_vtid)(struct scap_engine_handle engine, uint64_t tid, int64_t *vtid);
+
+	/**
+	 * @brief get the current process id in the init pid namespace
+	 * @param engine wraps the pointer to the engine-specific handle
+	 * @param pid output parameter, pointer to the pid
+	 * @param error a SCAP_LASTERR_SIZE buffer for error messages
+	 * @return SCAP_SUCCESS or a failure code
+	 */
+	int32_t (*getpid_global)(struct scap_engine_handle engine, int64_t* pid, char* error);
 };
 
 struct scap_linux_platform
