+++ userspace/libscap/scap-int.h
@@ -100,7 +100,6 @@ int32_t scap_check_suppressed(struct scap_suppress *suppress, scap_evt *pevent,
 			      bool *suppressed, char *error);
 
 int32_t scap_procfs_get_threadlist(struct scap_engine_handle engine, struct ppm_proclist_info **procinfo_p, char *lasterr);
-int32_t scap_os_getpid_global(struct scap_engine_handle engine, int64_t *pid, char* error);
 
 //
 // Retrieve agent info.
