+++ userspace/libscap/scap_platform_api.c
@@ -15,6 +15,8 @@ limitations under the License.
 
 */
 
+#include <stdio.h>
+
 #include "scap_platform_api.h"
 #include "scap_platform_impl.h"
 
@@ -98,4 +100,16 @@ bool scap_is_thread_alive(scap_t* handle, int64_t pid, int64_t tid, const char* 
 
 	// keep on the safe side, don't consider threads dead too early
 	return true;
+}
+
+int32_t scap_getpid_global(scap_t* handle, int64_t* pid)
+{
+	if (handle && handle->m_platform && handle->m_platform->m_vtable->get_global_pid)
+	{
+		return handle->m_platform->m_vtable->get_global_pid(handle->m_platform, pid, handle->m_lasterr);
+	}
+
+	ASSERT(false);
+	snprintf(handle->m_lasterr, SCAP_LASTERR_SIZE, "Cannot get pid (capture not enabled)");
+	return SCAP_FAILURE;
 }
