+++ userspace/libscap/scap_platform_api.h
@@ -95,6 +95,9 @@ struct scap_threadinfo* scap_get_proc_table(struct scap* handle);
 // Check if the given thread exists in /proc
 bool scap_is_thread_alive(struct scap* handle, int64_t pid, int64_t tid, const char* comm);
 
+// like getpid() but returns the global PID even inside a container
+int32_t scap_getpid_global(struct scap* handle, int64_t* pid);
+
 #ifdef __cplusplus
 };
 #endif
