+++ userspace/libscap/scap_platform_impl.h
@@ -65,6 +65,7 @@ struct scap_platform_vtable
 
 	int32_t (*refresh_proc_table)(struct scap_platform*, struct scap_proclist* proclist);
 	bool (*is_thread_alive)(struct scap_platform*, int64_t pid, int64_t tid, const char* comm);
+	int32_t (*get_global_pid)(struct scap_platform*, int64_t *pid, char *error);
 
 	// close the platform structure
 	// clean up all data, make it ready for another call to `init_platform`
